reformat: put mongo logic from user-service.js into config/backend.js
